security: fix 5 additional vulnerabilities (#11-#15)
#11 Realloc overflow in redact_buffer initial alloc (SIZE_MAX/2 guard)
#12 Realloc overflow in redact_buffer growth (SIZE_MAX/2 guard)
#13 Type truncation in redact_batch return (INT_MAX guard)
#14 Integer overflow in parallel work distribution (size_t cast)
#15 Hardened path traversal prevention (component-level check)

+++ src/parallel.c
@@ -210,7 +210,8 @@ int parallel_process(ParallelCtx *ctx, const char **lines,
   }
 
   /* Distribute work */
-  size_t per_thread = (count + ctx->num_threads - 1) / ctx->num_threads;
+  size_t per_thread =
+      (count + (size_t)ctx->num_threads - 1) / (size_t)ctx->num_threads;
 
   for (int i = 0; i < ctx->num_threads; i++) {
     Worker *w = &ctx->workers[i];
@@ -218,7 +219,9 @@ int parallel_process(ParallelCtx *ctx, const char **lines,
     w->lengths = lengths;
     w->outputs = outputs;
     w->out_lengths = out_lengths;
-    w->start_idx = i * per_thread;
+    /* SECURITY FIX #14: Cast i to size_t before multiplying to prevent
+     * integer overflow when per_thread is large. */
+    w->start_idx = (size_t)i * per_thread;
     w->end_idx = w->start_idx + per_thread;
     if (w->end_idx > count)
       w->end_idx = count;
